Updated code to use vite and include a mui modal
