+++ dev-docker-compose.yml
@@ -1,0 +1,9 @@
+version: '3'
+
+services:
+  frontend:
+    build:
+      context: ./
+      dockerfile: dev.dockerfile
+    ports:
+      - 3000:3000
