+++ dev.dockerfile
@@ -1,0 +1,20 @@
+# Use an official Node.js runtime as a parent image
+FROM node:18-alpine
+
+# Set the working directory in the container
+WORKDIR /usr/src/app
+
+# Copy package.json and package-lock.json to the working directory
+COPY package*.json ./
+
+# Install dependencies
+RUN npm install
+
+# Copy the rest of the application code to the working directory
+COPY . .
+
+# Expose port 3000 to the outside world
+EXPOSE 3000
+
+# Define the command to run the app when the container starts
+CMD ["npm", "start"]
