+++ index.html
@@ -1,0 +1,7 @@
+<!-- index.html -->
+<body>
+  <noscript>You need to enable JavaScript to run this app.</noscript>
+  <div id="root"></div>
+
+  <script type="module" src="/src/index.jsx"></script>
+</body>
